Update to patina v17.0.0 (#79)
## Description

Updates to the latest Patina release integrating breaking changes.

The patch version of patina-dxe-core has been incremented to 2.1.2 to
reflect that no specific work is expected in QEMU platform integration
for this release.

---

Note: Waiting on the v17.0.0 release (in
https://github.com/OpenDevicePartnership/patina/pull/1150) to be
completed so this PR can pick up the new crates. It has been tested
locally against those changes. It will be moved out of draft after that
release is made.

---

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- QEMU boot to EFI shell

## Integration Instructions

- N/A

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>